Move tests where they belong
